Implemented a basic slide builder, with additional logic and error handling when starting and viewing a presentation
