+++ views/presentation-starter.pug
@@ -12,6 +12,8 @@ body
     div.container-fluid
         div.row
             div.col-lg-6.offset-lg-3.col-sm-8.offset-sm-2
+                if messages.error
+                    div.alert.alert-danger= messages.error
                 form(method="post" action="/present").mx-auto
                     div.form-group
                         input(type="text" name="presentationID" placeholder="enter a presentation id").form-control.form-control-lg
