--- views/slides.pug
@@ -1,9 +1,0 @@
-section Slide 1
-
-section Slide 2
-
-section Slide 3
-
-section Slide 4
-
-section Slide 5
